mcp/client/requirements.txt
mcp>=1.27.0
openai>=1.0.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
